+++ sim/sim.c
@@ -82,7 +82,7 @@ int sim_run(struct sim *s, sim_feed_ptr feed) {
 }
 
 int sim_open_position(struct sim *s, struct timeline *t, 
-		      position_t type, int n) {
+		      position_t type, double n) {
   
   struct position *p;
   if(position_alloc(p, t, type, n)){
